backend/requirements.txt
fastapi==0.115.12
# Plain uvicorn avoids uvicorn[standard] pulling watchfiles, which often needs Rust wheels on Windows/Python 3.14.
uvicorn==0.34.0
# 3.11.x is yanked / may build from source on Python 3.14; use a release with wheels.
aiohttp>=3.12,<4
prometheus-client==0.21.1
# Use a flexible range so pip can pick pydantic-core wheels for your Python (3.14 needs newer core than 2.33).
pydantic>=2.11.1,<3
playwright==1.52.0
bcrypt>=4.0.1
python-jose[cryptography]==3.3.0
